requirements.txt
# requirements-graphmind[base].txt
# GraphMind Chat Base
python-dotenv
tqdm==4.66.5
shortuuid==1.0.13
pydantic>=2.8.2, <3.0.0
six==1.16.0
oss2==2.18.4
langchain==0.3.0
langgraph==0.2.28
langchain-community==0.3.0
langchain-core==0.3.0
langchain-openai==0.2.0
langchain-redis==0.1.0
redis==5.0.8
# GraphMind File IO
unstructured==0.15.1
markdown==3.6
langchain-text-splitters==0.3.0
# GraphMind File Structuring
zhipuai==2.1.4.20230814
neo4j==5.23.0
pandas==2.2.2
numpy==1.26.4
openpyxl==3.1.5
# requirements-graphmind[graphrag].txt
environs>=11.0.0,<12.0.0
datashaper>=0.0.49,<0.1.0
# GraphRAG Vector Stores
azure-search-documents>=11.4.0,<12.0.0
lancedb>=0.12.0,<0.13.0
# GraphRAG Event Loops
uvloop>=0.20.0,<0.21.0; platform_system != 'Windows'
nest-asyncio>=1.6.0,<2.0.0; platform_system == 'Windows'
# GraphRAG Async IO
aiolimiter>=1.1.0,<2.0.0
aiofiles>=24.1.0,<25.0.0
# GraphRAG LLM
openai>=1.37.1,<2.0.0
nltk==3.9.1
tiktoken>=0.7.0,<0.8.0
# GraphRAG Data-Sci
numba==0.60.0
graspologic>=3.4.1,<4.0.0
networkx>=3,<4.0.0
fastparquet>=2024.2.0,<2025.0.0
scipy==1.12.0
# GraphRAG Configuration
pyyaml>=6.0.2,<7.0.0
pyaml-env>=1.2.1,<2.0.0
python-dotenv>=1.0.0,<2.0.0
# GraphRAG Network
tenacity>=8.1.0, <9.0.0, !=8.4.0    # 对齐 langchain 要求
swifter>=1.4.0,<2.0.0
pydantic>=2,<3.0.0
rich>=13.6.0,<14.0.0
textual>=0.78.0,<0.79.0
devtools>=0.12.2,<0.13.0
typing-extensions>=4.12.2,<5.0.0
# GraphRAG Azure
azure-storage-blob>=12.22.0,<13.0.0
azure-identity>=1.17.1,<2.0.0
json-repair>=0.28.4,<0.29.0
future>=1.0.0,<2.0.0
# GraphRAG Config Template
jinja2==3.1.4
# requirements-graphmind[grpc].txt
protobuf==5.27.2
grpcio==1.66.1
grpcio-tools==1.66.1
# requirements-graphmind[http].txt
fastapi==0.115.0
uvicorn==0.30.6
# requirements-graphmind[magic-pdf[full]].txt
# MinerU Base
magic-pdf[full]
# MinerU Model
# https://github.com/opendatalab/MinerU/blob/master/docs/how_to_download_models_en.md
# MinerU config at ~/magic-pdf.json
# {
#    "bucket_info":{
#        "bucket-name-1":["ak", "sk", "endpoint"],
#        "bucket-name-2":["ak", "sk", "endpoint"]
#    },
#    "models-dir":"/tmp/models",    # 这里要改成实际的模型存放路径
#    "device-mode":"cpu",           # cpu or gpu
#    "table-config": {
#        "model": "TableMaster",
#        "is_table_recog_enable": false,
#        "max_time": 400
#    }
#}
